Move ofa_xception.py to contrib/classification/ofa/networks/ package.
